backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
dnspython==2.4.2  # For MongoDB+SRV connection strings

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# Original ML Dependencies
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
shap==0.43.0

# Pydantic
pydantic[email]==2.5.0

# Logging
logging==0.4.9.6

reportlab==4.0.7